fix bug

+++ MyDouban/Categories/NSDate+Convenience.h
@@ -10,14 +10,14 @@
 
 @interface NSDate (Convenience)
 
--(int)year;
--(int)month;
--(int)day;
--(int)hour;
+-(NSInteger)year;
+-(NSInteger)month;
+-(NSInteger)day;
+-(NSInteger)hour;
 - (NSString*)weekString;
 -(NSDate *)offsetDay:(int)numDays;
 +(NSDate *)dateStartOfDay:(NSDate *)date;
-+ (int)dayBetweenStartDate:(NSDate *)startDate endDate:(NSDate *)endDate;
++ (NSInteger)dayBetweenStartDate:(NSDate *)startDate endDate:(NSDate *)endDate;
 + (NSDate *)dateFromString:(NSString *)dateString format:(NSString *)format;
 + (NSString *)stringFromDate:(NSDate *)date format:(NSString *)format;
